mixins: Implement MetadataHandlerMixin
Remove duplicate implementations of parsing in both astprocd and astwifid
