+++ astoria/common/mixins/__init__.py
@@ -1,5 +1,6 @@
 """Manager Mixins."""
 
 from .disk_handler import DiskHandlerMixin
+from .metadata_handler import MetadataHandlerMixin
 
-__all__ = ["DiskHandlerMixin"]
+__all__ = ["DiskHandlerMixin", "MetadataHandlerMixin"]
